+++ packages/kit/test/apps/basics/test/server.test.js
@@ -295,6 +295,17 @@ test.describe('Endpoints', () => {
 		expect(response.headers()['digest']).toEqual(`sha-256=${digest}`);
 	});
 
+	test('stream can be canceled with TypeError', async ({ request }) => {
+		const responseBefore = await request.get('/endpoint-output/stream-typeerror?what');
+		expect(await responseBefore.text()).toEqual('null');
+
+		const interruptedResponse = request.get('/endpoint-output/stream-typeerror');
+		await expect(interruptedResponse).rejects.toThrow('socket hang up');
+
+		const responseAfter = await request.get('/endpoint-output/stream-typeerror?what');
+		expect(await responseAfter.text()).toEqual('TypeError');
+	});
+
 	test('request body can be read slow', async ({ request }) => {
 		const data = randomBytes(1024 * 256);
 		const digest = createHash('sha256').update(data).digest('base64url');
@@ -379,6 +390,19 @@ test.describe('Errors', () => {
 		} else {
 			expect(/** @type {Response} */ (response).status()).toBe(400);
 		}
+	});
+
+	test('stack traces are not fixed twice', async ({ page }) => {
+		await page.goto('/errors/stack-trace');
+		expect(await page.textContent('#message')).toBe(
+			'This is your custom error page saying: "Cannot read properties of undefined (reading \'toUpperCase\')"'
+		);
+
+		// check the stack wasn't mutated
+		await page.goto('/errors/stack-trace');
+		expect(await page.textContent('#message')).toBe(
+			'This is your custom error page saying: "Cannot read properties of undefined (reading \'toUpperCase\')"'
+		);
 	});
 });