+++ packages/kit/test/prerendering/basics/src/hooks.js
@@ -2,13 +2,14 @@ import { prerendering } from '$app/env';
 
 const initial_prerendering = prerendering;
 
+/** @type {import('@sveltejs/kit').Handle} */
 export const handle = async ({ event, resolve }) => {
 	if (event.url.pathname === '/prerendering-true' && prerendering) {
 		return await resolve(event, {
-			transformPage: ({ html }) =>
+			transformPageChunk: ({ html }) =>
 				html
-					.replace('__INITIAL_PRERENDERING__', initial_prerendering)
-					.replace('__PRERENDERING__', prerendering)
+					.replace('__INITIAL_PRERENDERING__', String(initial_prerendering))
+					.replace('__PRERENDERING__', String(prerendering))
 		});
 	}
 	return await resolve(event);